pyproject.toml
[tool.poetry]
name = "tp_python_code"
version = "1.0.0"
description = ""
authors = ["Mathias EMON <m.emon@myskolae.fr>"]
readme = "README.md"
packages = [
  { include = "src" },
  { include = "tp1", from = "src" },
  { include = "tp2", from = "src" },
  { include = "tp3", from = "src" },
  { include = "tp4", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
scapy = "^2.6.0"
pygal = "^3.0.5"
fpdf2 = "^2.8.1"
pytest = "^8.3.5"
reportlab = "^4.4.9"
matplotlib = "^3.10.8"
capstone = "^5.0.1"
pytesseract = "^0.3.10"
pillow = "^12.1.1"
pwntools = "^4.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tp1 = "tp1.main:main"
tp2 = "tp2.main:main"
tp3 = "tp3.main:main"
tp4 = "tp4.main:main"